add top trending words card
