+++ src/components/SignalCard/SignalCardWrapper.js
@@ -2,7 +2,7 @@ import React from 'react'
 import cx from 'classnames'
 import MultilineText from '../MultilineText/MultilineText'
 import { SignalTypeIcon } from './controls/SignalControls'
-import styles from './SignalCard.module.scss'
+import styles from './card/SignalCard.module.scss'
 
 export const SignalCardWrapper = ({
   isModal = true,
