+++ src/components/SignalCard/SignalCardsGrid.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { push } from 'react-router-redux'
 import cx from 'classnames'
 import { connect } from 'react-redux'
-import SignalCard from './SignalCard'
+import SignalCard from './card/SignalCard'
 import {
   removeTrigger,
   toggleTrigger
