+++ src/components/SignalCard/card/SignalCard.module.scss
@@ -1,4 +1,4 @@
-@import "~@santiment-network/ui/mixins.scss";
+@import "~@santiment-network/ui/mixins";
 
 .wrapper {
   display: flex;
