+++ src/ducks/Signals/signalFormManager/sharedForm/SharedTriggerForm.js
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react'
 import Button from '@santiment-network/ui/Button'
 import { couldShowChart, mapFormPropsToTrigger } from '../../utils/utils'
 import SignalPreview from '../../chart/SignalPreview'
-import SignalCard from '../../../../components/SignalCard/SignalCard'
+import SignalCard from '../../../../components/SignalCard/card/SignalCard'
 import NoSignalPreview from '../../chart/NoSignalPreview'
 import styles from './ShareTriggerForm.module.scss'
 import { fetchHistorySignalPoints } from '../../common/actions'
