+++ src/queries/FeedGQL.js
@@ -1,5 +1,6 @@
 import gql from 'graphql-tag'
 import { INSIGHT_COMMON_FRAGMENT } from './InsightsGQL'
+import { TRIGGERS_COMMON_FRAGMENT } from '../ducks/Signals/common/queries'
 
 export const FEED_QUERY = gql`
   query timelineEvents($limit: Int, $cursor: CursorInput) {
@@ -18,9 +19,13 @@ export const FEED_QUERY = gql`
         post {
           ...insightCommon
         }
+        trigger {
+          ...triggersCommon
+        }
         __typename
       }
     }
   }
   ${INSIGHT_COMMON_FRAGMENT}
+  ${TRIGGERS_COMMON_FRAGMENT}
 `
